Added watchdog
HAL must kick watchdog regularly, or the watchdog will cause a the
processor to restart.
The board now sends a message on startup, allowing the user to know when
the board is reset.
A software reset command may be used to force the board to restart. The
reset is caused by locking up, and not kicking the watchdog.